+++ src/application/Event.java
@@ -1,9 +1,6 @@
-package application.TimeUnits;
+package application;
 
 import java.util.Date;
-
-import application.Exceptions.InvalidColourException;
-import application.Exceptions.NullEventEndPointException;
 /**
  * Base class for all user created events. Should not be constructed manually.
  */
