+++ src/application/InvalidColourException.java
@@ -1,4 +1,4 @@
-package application.Exceptions;
+package application;
 
 public class InvalidColourException extends Exception {
