+++ src/application/InvalidPasswordException.java
@@ -1,4 +1,4 @@
-package application.Exceptions;
+package application;
 
 public class InvalidPasswordException extends Exception {
