+++ src/application/InvalidUsernameException.java
@@ -1,4 +1,4 @@
-package application.Exceptions;
+package application;
 
 public class InvalidUsernameException extends Exception {
