+++ src/application/NullEventEndPointException.java
@@ -1,4 +1,4 @@
-package application.Exceptions;
+package application;
 
 public class NullEventEndPointException extends Exception {
