+++ src/application/TimeUnit.java
@@ -1,8 +1,6 @@
-package application.TimeUnits;
+package application;
 
 import java.util.Date;
-
-import application.Exceptions.NullEventEndPointException;
 
 /**
  *Base class for all time events, should never be constructed directly.
