+++ src/application/Week.java
@@ -1,11 +1,6 @@
-package application.TimeUnits;
+package application;
 
 import java.util.Date;
-
-import application.Exceptions.EventOutsideTimeUnitException;
-import application.Exceptions.InvalidColourException;
-import application.Exceptions.NullEventEndPointException;
-
 import java.util.Calendar;
 
 public class Week extends TimeUnit{
